partially working decoupling of network code to ExchangeApi
